+++ src/Classifiers/LogitBoost.php
@@ -126,7 +126,7 @@ class LogitBoost implements Estimator, Learner, Probabilistic, RanksFeatures, Ve
     protected int $evalInterval;
 
     /**
-     * The number of epochs without improvement in the validation score to wait before considering an early stop.
+     * The number of evaluations without improvement in the validation score to wait before considering an early stop.
      *
      * @var positive-int
      */
@@ -198,7 +198,7 @@ class LogitBoost implements Estimator, Learner, Probabilistic, RanksFeatures, Ve
         float $rate = 0.1,
         float $ratio = 0.5,
         int $epochs = 1000,
-        float $minChange = 1e-4,
+        float $minChange = 1e-5,
         int $evalInterval = 3,
         int $window = 5,
         float $holdOut = 0.1,