+++ src/Classifiers/SoftmaxClassifier.php
@@ -116,7 +116,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
     protected int $evalInterval;
 
     /**
-     * The number of epochs without improvement in the validation score to wait before considering an early stop.
+     * The number of evaluations without improvement in the validation score to wait before considering an early stop.
      *
      * @var positive-int
      */
@@ -198,9 +198,9 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
         float $l1Penalty = 1e-4,
         float $l2Penalty = 1e-4,
         int $epochs = 1000,
-        float $minChange = 1e-4,
-        int $evalInterval = 3,
-        int $window = 5,
+        float $minChange = 1e-5,
+        int $evalInterval = 1,
+        int $window = 10,
         float $holdOut = 0.1,
         ?ClassificationLoss $costFn = null,
         ?Metric $metric = null